pyproject.toml
[project]
name = "llm-collector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.118.0",
    "modal>=1.1.4",
    "openai>=1.109.1",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
    "supabase>=2.20.0",
]
